# backend/openapi.yml
openapi: 3.0.3
info:
  title: UpChatz API
  version: 1.0.0
  description: Multi-tenant WhatsApp Business Platform (Cloud API) for Indian businesses.
paths:
...
  /api/v1/catalog/collections/:
    get:
      operationId: catalog_collections_list
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCollectionList'
          description: ''
    post:
      operationId: catalog_collections_create
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CollectionWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CollectionWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: ''
  /api/v1/catalog/collections/{id}/:
    get:
      operationId: catalog_collections_retrieve
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: ''
    patch:
      operationId: catalog_collections_partial_update
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCollectionWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCollectionWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCollectionWriteRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: ''
    delete:
      operationId: catalog_collections_destroy
      description: Its products move out of the collection.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/catalog/collections/reorder/:
    post:
      operationId: catalog_collections_reorder_create
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/catalog/meta-catalogs/:
    get:
      operationId: catalog_meta_catalogs_list
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMetaCatalogList'
          description: ''
    post:
      operationId: catalog_meta_catalogs_create
      description: 409 catalog_permissions_missing when the token lacks catalog permissions.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetaCatalogConnectRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MetaCatalogConnectRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MetaCatalogConnectRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaCatalog'
          description: ''
  /api/v1/catalog/meta-catalogs/{id}/:
    get:
      operationId: catalog_meta_catalogs_retrieve
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaCatalog'
          description: ''
    delete:
      operationId: catalog_meta_catalogs_destroy
      description: Disconnects locally; the store falls back to bot mode.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/catalog/meta-catalogs/{id}/commerce-settings/:
    patch:
      operationId: catalog_meta_catalogs_commerce_settings_partial_update
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCommerceSettingsRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCommerceSettingsRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCommerceSettingsRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaCatalog'
          description: ''
  /api/v1/catalog/meta-catalogs/{id}/sync/:
    post:
      operationId: catalog_meta_catalogs_sync_create
      description: Queue a full resync.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaCatalog'
          description: ''
  /api/v1/catalog/meta-catalogs/available/:
    get:
      operationId: catalog_meta_catalogs_available_list
      description: Catalogs owned by the seller's Meta business (unpaginated).
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: query
        name: waba_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableCatalog'
          description: ''
  /api/v1/catalog/products/:
    get:
      operationId: catalog_products_list
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: query
        name: availability
        schema:
          type: string
          enum:
          - in_stock
          - out_of_stock
      - in: query
        name: collection
        schema:
          type: string
        description: Collection id, or none.
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - in: query
        name: is_active
        schema:
          type: boolean
      - in: query
        name: meta_review_status
        schema:
          type: string
          enum:
          - approved
          - none
          - outdated
          - pending
          - rejected
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
        description: Name or SKU.
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductList'
          description: ''
    post:
      operationId: catalog_products_create
      description: '409-free: a duplicate SKU is a 400 invalid error on sku.'
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
  /api/v1/catalog/products/{id}/:
    get:
      operationId: catalog_products_retrieve
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
    patch:
      operationId: catalog_products_partial_update
      description: sku is read-only after create.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProductWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProductWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProductWriteRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
    delete:
      operationId: catalog_products_destroy
      description: Queues a Meta catalog delete; past order items keep their snapshots.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/catalog/products/{id}/image/:
    post:
      operationId: catalog_products_image_create
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductImageUploadRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
    delete:
      operationId: catalog_products_image_destroy
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - catalog
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/catalog/products/import/:
    post:
      operationId: catalog_products_import_create
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      tags:
      - catalog
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductImportUploadRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductImportResult'
          description: ''
  /api/v1/catalog/products/reorder/:
    post:
      operationId: catalog_products_reorder_create
      description: Compose with DRF mixins (ListModelMixin, ...) for partial CRUD.
      parameters:
      - in: header
        name: X-Workspace-ID
        schema:
          type: string
          format: uuid
        description: Workspace to act in. The caller must be a member.
        required: true
      tags:
      - catalog
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    AvailableCatalog:
      type: object
      properties:
        id:
          type: string
          readOnly: true
          description: Meta catalog id.
        name:
          type: string
          readOnly: true
      required:
      - id
      - name
...
    Collection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
        description:
          type: string
          readOnly: true
        position:
          type: integer
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
        product_count:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - description
      - id
      - is_active
      - name
      - position
      - product_count
      - updated_at
...
    CollectionWriteRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Shown as a list row title.
          maxLength: 24
        description:
          type: string
          maxLength: 72
        position:
          type: integer
        is_active:
          type: boolean
      required:
      - name
...
    MetaCatalogConnectRequest:
      type: object
      description: Connect an existing catalog (``catalog_id``) or create one (``create_name``).
      properties:
        waba_id:
          type: string
          format: uuid
          description: Id of the WhatsAppBusinessAccount.
        catalog_id:
          type: string
          minLength: 1
          pattern: ^[0-9]{1,64}$
        create_name:
          type: string
          minLength: 1
          maxLength: 100
      required:
      - waba_id
...
    PaginatedCollectionList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Collection'
...
    PaginatedMetaCatalogList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/MetaCatalog'
...
    PatchedCollectionWriteRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Shown as a list row title.
          maxLength: 24
        description:
          type: string
          maxLength: 72
        position:
          type: integer
        is_active:
          type: boolean
    PatchedCommerceSettingsRequest:
      type: object
      description: |-
        Per-number cart and catalog visibility. As a PATCH body only ``phone_number_id`` and the
        booleans are read.
      properties:
        phone_number_id:
          type: string
          format: uuid
        is_cart_enabled:
          type: boolean
        is_catalog_visible:
          type: boolean
...
    ReorderRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
            format: uuid
          description: The full ordering; positions are rewritten 0..n.
          maxItems: 5000
      required:
      - ids
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT